refactor and separate classes
