+++ src/main/java/bio/guoda/preston/index/SearchIndex.java
@@ -1,4 +1,10 @@
 package bio.guoda.preston.index;
 
-public interface SearchIndex extends SearchIndexReadable, SearchIndexWritable {
+import org.apache.lucene.document.Document;
+
+import java.io.IOException;
+
+public interface SearchIndex extends SearchIndexReadOnly {
+
+    void put(Document doc) throws IOException;
 }
